refactor: count 쿼리의 반환값을 int에서 long으로 수정

+++ src/main/java/aegis/server/domain/study/dto/response/StudySummaryResponse.java
@@ -8,8 +8,8 @@ public record StudySummaryResponse(
         String title,
         StudyCategory category,
         StudyLevel level,
-        Integer participantCount,
-        Integer maxParticipants,
+        long participantCount,
+        int maxParticipants,
         String schedule,
         String instructor) {
 
@@ -18,8 +18,8 @@ public record StudySummaryResponse(
             String title,
             StudyCategory category,
             StudyLevel level,
-            Integer participantCount,
-            Integer maxParticipants,
+            long participantCount,
+            int maxParticipants,
             String schedule,
             String instructor) {
         return new StudySummaryResponse(
